improve size

+++ contracts/interfaces/IFuulManager.sol
@@ -33,9 +33,9 @@ interface IFuulManager {
     }
 
     struct FeesInformation {
-        uint8 protocolFee;
-        uint8 attributorFee;
-        uint8 clientFee;
+        uint256 protocolFee;
+        uint256 attributorFee;
+        uint256 clientFee;
         address protocolFeeCollector;
         uint256 nftFixedFeeAmount;
         address nftFeeCurrency;
@@ -63,15 +63,15 @@ interface IFuulManager {
         address currency
     ) external view returns (uint256);
 
-    function protocolFee() external view returns (uint8 fees);
+    function protocolFee() external view returns (uint256 fees);
 
     function protocolFeeCollector() external view returns (address);
 
     function getFeesInformation() external returns (FeesInformation memory);
 
-    function clientFee() external view returns (uint8 fees);
+    function clientFee() external view returns (uint256 fees);
 
-    function attributorFee() external view returns (uint8 fees);
+    function attributorFee() external view returns (uint256 fees);
 
     function nftFeeCurrency() external view returns (address);
 
@@ -87,11 +87,11 @@ interface IFuulManager {
       ║        FEES VARIABLES       ║
       ╚═════════════════════════════╝*/
 
-    function setProtocolFee(uint8 value) external;
+    function setProtocolFee(uint256 value) external;
 
-    function setClientFee(uint8 value) external;
+    function setClientFee(uint256 value) external;
 
-    function setAttributorFee(uint8 value) external;
+    function setAttributorFee(uint256 value) external;
 
     /*╔═════════════════════════════╗
       ║       TOKEN CURRENCIES      ║
